Merged PR 3: Certificate event + tasks 72-73
This pull request contains the changes to the ID contract that allow a newly created certificate to be returned through an event (merged from branch Jonathan/events) as well as the changes from tasks 72 and 73:

- Making ID.addCertificate(Attribute, Certificate) only callable by the ID's owner
- Creating functions IDController.addCertificate(Attribute, Certificate) and IDController.addCertificate(bytes32, Certificate).

and the necessary modifications to the unit tests (which all pass). The failing IDController.deleteID test has been temporarily commented out.

Related work items: #73, #72